Changements dus à destinRevolution entre autres

+++ glisseur.cpp
@@ -13,16 +13,16 @@ Glisseur::Glisseur(QJsonObject json)
             "Il manque un param"
             );
 
-    m_Minimum = json["valeur_min"].toString().toInt();
-    m_Maximum = json["valeur_max"].toString().toInt();
-    m_ValeurDepart = json["valeur_depart"].toString().toInt();
+    m_Minimum = json["valeur_min"].toString().toDouble();
+    m_Maximum = json["valeur_max"].toString().toDouble();
+    m_ValeurDepart = json["valeur_depart"].toString().toDouble();
     m_IdCaracAssociee = json["carac_id"].toString();
 }
 
 Glisseur::Glisseur(QString valeur_min, QString valeur_max, QString valeur_depart, QString carac_id)
 {
-    m_Minimum = valeur_min.toInt();
-    m_Maximum = valeur_max.toInt();
-    m_ValeurDepart = valeur_depart.toInt();
+    m_Minimum = valeur_min.toDouble();
+    m_Maximum = valeur_max.toDouble();
+    m_ValeurDepart = valeur_depart.toDouble();
     m_IdCaracAssociee = carac_id;
 }
